Refactored flow into FlowProcessor
