+++ packages/appengine/src/mixins/createInquire.ts
@@ -1,9 +1,7 @@
 import { baseProvisionerType } from '../index'
 import { parser } from '../parser'
 import { AppEngineState, AppManifest, AppObject, Helper } from '../appObject'
-import { Flow } from '../contracts'
-import { skippedSteps as testSteps } from './flowSamples'
-import chalk from 'chalk'
+import { FlowProcessor, skippedSteps as testSteps } from '../flow'
 import { inspect } from 'util'
 import createDebug from 'debug'
 
@@ -11,100 +9,16 @@ const debug = createDebug('@appengine:createInquire')
 
 export const createInquireMixin = (base: baseProvisionerType) => class extends base {
 
-    inquireResponses: any = {}
-    inquireExtensions = new Map<string, Flow.c6oExtensions>()
-
     async createInquire(args) {
+        // Steps will come from the applicationSpec but for now, we use test data
         // const steps = this.manager.document.spec?.provisioner?.steps
-        const steps = testSteps
-        for (const step of Flow.each(steps))
-            await this.processStep(step)
+
+        const flowProcessor = new FlowProcessor(this.manager.inquirer, this.manager.document)
+        await flowProcessor.process(testSteps)
 
         // Stop provisioning from going any further
-        throw new Error(`NOT IMPLEMENTED ${inspect(this.inquireResponses)}`)
+        throw new Error(`NOT IMPLEMENTED ${inspect(flowProcessor.inquireResponses)}`)
     }
-
-    async processStep(step: Flow.step) {
-        debug('Processing step %s', step.name)
-
-        if (step.skip) {
-            debug('Processing skip function %o', step.skip)
-            const skipFunction =  new Function('answers',  step.skip)
-            const skip = skipFunction.call(this.manager.document, this.inquireResponses)
-            debug('Skip function says %s', skip)
-            if (skip)
-                return
-        }
-
-        if (step.sections)
-            for (const section of step.sections)
-                await this.processSection(section)
-        else if (step.inquire)
-            await this.processInquire(step.inquire)
-        else
-            throw new Error(`Step ${step.name} lacks any sections or inquire`)
-    }
-
-    async processSection(section: Flow.section) {
-        debug('Processing section %s', section.title)
-
-        // We can't use console.log because vorpal does something
-        // to console.log so we write directly to stdout
-        process.stdout.write(chalk.yellowBright.bold('\n❯ Section: ') + chalk.yellowBright(section.title) + '\n\n')
-        await this.processInquire(section.inquire)
-    }
-
-    async processInquire(inquireField: Flow.inquireType) {
-        const inquireFields = Array.isArray(inquireField) ?
-            inquireField :
-            [inquireField]
-
-        const asks = inquireFields.map(item => this.mapInquire(item))
-        const responses = await this.manager.inquirer.prompt(asks)
-
-        // Add to the responses bag
-        this.inquireResponses = {
-            ...this.inquireResponses,
-            ...responses
-        }
-    }
-
-    /**
-     * Maps from a Flow.inquireType item to inquire.prompt items
-     * @param inquireItem
-     */
-    mapInquire(inquireItem: Flow.inquireStep) {
-        debug('Mapping from %o', inquireItem)
-
-        // Inquire is forgiving and ignores fields it doesn't understand
-        // but let's be good and strip out c6o
-        const { c6o, ...rest } = inquireItem
-
-        if (Flow.isFunctionString(rest.validate))
-            rest.validate = new Function('value', 'answers', rest.validate).bind(this.manager.document)
-
-        if (Flow.isFunctionString(rest.when))
-            rest.when = new Function('answers', rest.when).bind(this.manager.document)
-
-        if (rest.type === 'password' && !rest.mask)
-            rest.mask = '*'
-
-        // For choices, we have to convert separator strings to the separator object
-        if (rest.choices)
-            rest.choices = rest.choices.map(choice =>
-                choice === Flow.choiceSeparator ?
-                    new this.manager.inquirer.Separator() :
-                    choice
-            )
-
-        // Stash the extensions
-        this.inquireExtensions.set(rest.name, c6o)
-
-        debug('Mapped to %o', rest)
-        return rest
-    }
-
-
 
     helper = new Helper()
     async oldCreateInquire(args) {
